Refactor CookieExpirationHandler to implement AuthenticationRefreshHandler. Save the authenticators in cookie SignIn and add Tolerance to the cookie expiration
